+++ class_names.py
@@ -1,0 +1,27 @@
+in_selected_classes = [
+    "ladybug",
+    "hamster",
+    "zebra",
+    "flamingo",
+    "forklift",
+    "monarch",
+    "tiger",
+    "fountain",
+    "knot",
+    "koala",
+    "castle",
+    "candle",
+    "hummingbird",
+    "snail",
+    "coyote",
+    "barn",
+    "strawberry",
+    "chihuahua",
+    "broccoli",
+    "starfish",
+    "jellyfish",
+    "microphone",
+    "gorilla",
+    "hare",
+    "lion",
+]
