+++ in100_class_index.json
@@ -1,0 +1,378 @@
+{
+    "stingray": [
+        6,
+        "n01498041"
+    ],
+    "hen": [
+        8,
+        "n01514859"
+    ],
+    "magpie": [
+        18,
+        "n01582220"
+    ],
+    "kite": [
+        21,
+        "n01608432"
+    ],
+    "vulture": [
+        23,
+        "n01616318"
+    ],
+    "agama": [
+        42,
+        "n01687978"
+    ],
+    "tick": [
+        78,
+        "n01776313"
+    ],
+    "quail": [
+        85,
+        "n01806567"
+    ],
+    "hummingbird": [
+        94,
+        "n01833805"
+    ],
+    "koala": [
+        105,
+        "n01882714"
+    ],
+    "jellyfish": [
+        107,
+        "n01910747"
+    ],
+    "snail": [
+        113,
+        "n01944390"
+    ],
+    "flamingo": [
+        130,
+        "n02007558"
+    ],
+    "chihuahua": [
+        151,
+        "n02085620"
+    ],
+    "coyote": [
+        272,
+        "n02114855"
+    ],
+    "tabby": [
+        281,
+        "n02123045"
+    ],
+    "leopard": [
+        288,
+        "n02128385"
+    ],
+    "lion": [
+        291,
+        "n02129165"
+    ],
+    "tiger": [
+        292,
+        "n02129604"
+    ],
+    "ladybug": [
+        301,
+        "n02165456"
+    ],
+    "fly": [
+        308,
+        "n02190166"
+    ],
+    "ant": [
+        310,
+        "n02219486"
+    ],
+    "grasshopper": [
+        311,
+        "n02226429"
+    ],
+    "monarch": [
+        323,
+        "n02279972"
+    ],
+    "starfish": [
+        327,
+        "n02317335"
+    ],
+    "hare": [
+        331,
+        "n02326432"
+    ],
+    "hamster": [
+        333,
+        "n02342885"
+    ],
+    "beaver": [
+        337,
+        "n02363005"
+    ],
+    "zebra": [
+        340,
+        "n02391049"
+    ],
+    "ox": [
+        345,
+        "n02403003"
+    ],
+    "impala": [
+        352,
+        "n02422699"
+    ],
+    "mink": [
+        357,
+        "n02442845"
+    ],
+    "otter": [
+        360,
+        "n02444819"
+    ],
+    "gorilla": [
+        366,
+        "n02480855"
+    ],
+    "sturgeon": [
+        394,
+        "n02640242"
+    ],
+    "accordion": [
+        401,
+        "n02672831"
+    ],
+    "ambulance": [
+        407,
+        "n02701002"
+    ],
+    "apron": [
+        411,
+        "n02730930"
+    ],
+    "backpack": [
+        414,
+        "n02769748"
+    ],
+    "balloon": [
+        417,
+        "n02782093"
+    ],
+    "banjo": [
+        420,
+        "n02787622"
+    ],
+    "barn": [
+        425,
+        "n02793495"
+    ],
+    "baseball": [
+        429,
+        "n02799071"
+    ],
+    "basketball": [
+        430,
+        "n02802426"
+    ],
+    "beacon": [
+        437,
+        "n02814860"
+    ],
+    "binder": [
+        446,
+        "n02840245"
+    ],
+    "broom": [
+        462,
+        "n02906734"
+    ],
+    "candle": [
+        470,
+        "n02948072"
+    ],
+    "castle": [
+        483,
+        "n02980441"
+    ],
+    "chain": [
+        488,
+        "n02999410"
+    ],
+    "chest": [
+        492,
+        "n03014705"
+    ],
+    "church": [
+        497,
+        "n03028079"
+    ],
+    "cinema": [
+        498,
+        "n03032252"
+    ],
+    "cradle": [
+        516,
+        "n03125729"
+    ],
+    "dam": [
+        525,
+        "n03160309"
+    ],
+    "desk": [
+        526,
+        "n03179701"
+    ],
+    "dome": [
+        538,
+        "n03220513"
+    ],
+    "drum": [
+        541,
+        "n03249569"
+    ],
+    "envelope": [
+        549,
+        "n03291819"
+    ],
+    "forklift": [
+        561,
+        "n03384352"
+    ],
+    "fountain": [
+        562,
+        "n03388043"
+    ],
+    "gown": [
+        578,
+        "n03450230"
+    ],
+    "hammer": [
+        587,
+        "n03481172"
+    ],
+    "jean": [
+        608,
+        "n03594734"
+    ],
+    "jeep": [
+        609,
+        "n03594945"
+    ],
+    "knot": [
+        616,
+        "n03627232"
+    ],
+    "laptop": [
+        620,
+        "n03642806"
+    ],
+    "library": [
+        624,
+        "n03661043"
+    ],
+    "lipstick": [
+        629,
+        "n03676483"
+    ],
+    "mask": [
+        643,
+        "n03724870"
+    ],
+    "maze": [
+        646,
+        "n03733281"
+    ],
+    "microphone": [
+        650,
+        "n03759954"
+    ],
+    "microwave": [
+        651,
+        "n03761084"
+    ],
+    "missile": [
+        657,
+        "n03773504"
+    ],
+    "nail": [
+        677,
+        "n03804744"
+    ],
+    "perfume": [
+        711,
+        "n03916031"
+    ],
+    "pillow": [
+        721,
+        "n03938244"
+    ],
+    "printer": [
+        742,
+        "n04004767"
+    ],
+    "purse": [
+        748,
+        "n04026417"
+    ],
+    "rifle": [
+        764,
+        "n04090263"
+    ],
+    "sandal": [
+        774,
+        "n04133789"
+    ],
+    "screw": [
+        783,
+        "n04153751"
+    ],
+    "stage": [
+        819,
+        "n04296562"
+    ],
+    "stove": [
+        827,
+        "n04330267"
+    ],
+    "swing": [
+        843,
+        "n04371774"
+    ],
+    "television": [
+        851,
+        "n04404412"
+    ],
+    "tractor": [
+        866,
+        "n04465501"
+    ],
+    "tripod": [
+        872,
+        "n04485082"
+    ],
+    "umbrella": [
+        879,
+        "n04507155"
+    ],
+    "violin": [
+        889,
+        "n04536866"
+    ],
+    "whistle": [
+        902,
+        "n04579432"
+    ],
+    "wreck": [
+        913,
+        "n04606251"
+    ],
+    "broccoli": [
+        937,
+        "n07714990"
+    ],
+    "strawberry": [
+        949,
+        "n07745940"
+    ]
+}
